Corrige errores de compilación en CI (icono Star sin importar y tipo Float en Slider)
Icons.Filled.Star es una propiedad de extensión declarada en otro paquete, por lo
que requiere import explícito en vez de referenciarse por nombre completamente
calificado. En ParameterSliderRow, el fallback del rango del Slider mezclaba
Double y Float, lo que el compilador no podía inferir como Float. De paso se
añaden índices a las foreign keys de UserBadgeEntity para evitar el warning de
Room sobre full table scans.

+++ app/src/main/java/com/kidslab/physicslab/ui/components/CompareExplainCard.kt
@@ -83,7 +83,7 @@ fun ParameterSliderRow(
         androidx.compose.material3.Slider(
             value = value.toFloat(),
             onValueChange = { onValueChange(it.toDouble()) },
-            valueRange = min.toFloat()..(if (max > min) max.toFloat() else (min + 1f)),
+            valueRange = min.toFloat()..(if (max > min) max.toFloat() else (min + 1.0).toFloat()),
             enabled = enabled && max > min
         )
     }
